test: cover reserved default dev identities
