+++ Packages/iOS/CmuxMobileShell/Tests/CmuxMobileShellTests/PairedMacBackupProtocolTests.swift
@@ -4,40 +4,6 @@ import Testing
 
 @testable import CmuxMobileShell
 
-private final class UnavailableV2PairedMacURLProtocol: URLProtocol, @unchecked Sendable {
-    nonisolated(unsafe) static let recorder = PairedMacBackupRequestRecorder()
-
-    override class func canInit(with request: URLRequest) -> Bool { true }
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest { request }
-
-    override func startLoading() {
-        Task { await Self.recorder.record(request) }
-        let response = HTTPURLResponse(
-            url: request.url!,
-            statusCode: 404,
-            httpVersion: nil,
-            headerFields: nil
-        )!
-        client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-        client?.urlProtocol(self, didLoad: Data(#"{"error":"not_found"}"#.utf8))
-        client?.urlProtocolDidFinishLoading(self)
-    }
-
-    override func stopLoading() {}
-}
-
-private actor PairedMacBackupRequestRecorder {
-    private(set) var requests: [URLRequest] = []
-
-    func record(_ request: URLRequest) {
-        requests.append(request)
-    }
-
-    func reset() {
-        requests = []
-    }
-}
-
 @Suite(.serialized) struct PairedMacBackupProtocolTests {
     private let strictScope = "cmux-dev:v2:ZmVhdHVyZQ"
