+++ Packages/iOS/CmuxMobileShell/Tests/CmuxMobileShellTests/PairedMacBackupRequestRecorder.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+actor PairedMacBackupRequestRecorder {
+    private(set) var requests: [URLRequest] = []
+
+    func record(_ request: URLRequest) {
+        requests.append(request)
+    }
+
+    func reset() {
+        requests = []
+    }
+}
